Extract tickTackToe main component to pages directory
